requirements.txt
crewai>=0.28.8
crewai-tools>=0.1.6
python-dotenv>=1.0.0
pydantic>=2.0.0
colorama>=0.4.6
rich>=13.0.0
pipedrive-python-lib>=1.0.0
openai>=1.37.0
PyPDF2>=3.0.0
python-docx>=1.0.0
requests>=2.31.0
python-dateutil>=2.9.0
